refactor: make minimal-elf non-internal package
